Add ChangeVelocityCommand

+++ include/Point.h
@@ -9,19 +9,19 @@ class Point : public IProperty
 {
 public:
     Point() = default;
-    Point(int x = 0, int y = 0)
+    Point(double x = 0, double y = 0)
         : m_x{x}
         , m_y{y}
     {}
 
     std::string name() override { return "position"; }
 
-    int getX() { return m_x; }
-    int getY() { return m_y; }
+    double getX() { return m_x; }
+    double getY() { return m_y; }
 
 private:
-    int m_x{0};
-    int m_y{0};
+    double m_x{0};
+    double m_y{0};
 };
 
-}
+}  // namespace otus
